feat: Various frontend enhancements
- Use same height for navbar buttons as in sidebar
- Fix different badge style in system info
- Fix missing image poll settings save button
- Move Webhook settings to new image settings tab
