+++ frontend/messages/de.json
@@ -375,7 +375,7 @@
 	"copyWebhookSecret": "Webhook-Secret kopieren",
 	"revealSecret": "Secret anzeigen",
 	"hideSecret": "Secret ausblenden",
-	"saveSecretNow": "Speichere dieses Secret jetzt - es wird nicht erneut angezeigt.",
+	"saveSecretNow": "Speichere dieses Secret jetzt. Es wird nicht erneut angezeigt.",
 	"docsLinkComingSoon": "TODO: Link zur Dokumentation",
 	"connectRepository": "Repository verbinden",
 	"updateRepository": "Repository aktualisieren",
