+++ frontend/messages/en.json
@@ -375,7 +375,7 @@
 	"copyWebhookSecret": "Copy webhook secret",
 	"revealSecret": "Reveal secret",
 	"hideSecret": "Hide secret",
-	"saveSecretNow": "Save this secret now - it won't be shown again.",
+	"saveSecretNow": "Save this secret now as it won't be shown again.",
 	"docsLinkComingSoon": "TODO: Link to docs",
 	"connectRepository": "Connect Repository",
 	"updateRepository": "Update Repository",
@@ -531,7 +531,7 @@
 	"regenerateWebhook": "Regenerate",
 	"revokeWebhook": "Revoke",
 	"imagePullWebhookSecretModalTitle": "Webhook Secret",
-	"imagePullWebhookSecretModalDescription": "Copy the secret now — it will not be shown again. Send it as Authorization: Bearer <secret>, or for Docker Hub append ?token=<secret> to the webhook URL.",
+	"imagePullWebhookSecretModalDescription": "Copy the secret now as it will not be shown again. Send it as Authorization: Bearer <secret>, or for Docker Hub append ?token=<secret> to the webhook URL.",
 	"imagePullWebhookUrl": "Webhook URL",
 	"imagePullWebhookSecret": "Secret",
 	"revokeWebhookConfirmDescription": "This will immediately invalidate the current webhook secret. Any existing integrations using the old secret will stop working.",
