+++ frontend/src/routes/_authenticated/applications/$id/details.tsx
@@ -1,21 +1,6 @@
-import {
-	deployApplication,
-	generateImageWebhook,
-	revokeImageWebhook,
-	HealthStatus,
-	SyncStatus,
-	type Application,
-} from "@/lib/applications";
+import { deployApplication, HealthStatus, SyncStatus, type Application } from "@/lib/applications";
 import { createFileRoute } from "@tanstack/react-router";
-import {
-	Clock,
-	ExternalLink,
-	GitBranch,
-	GitCommit,
-	RefreshCw,
-	Server,
-	Webhook,
-} from "lucide-react";
+import { Clock, ExternalLink, GitBranch, GitCommit, RefreshCw, Server } from "lucide-react";
 import { ApplicationStatusBadge } from "@/components/badges/application-status-badge";
 import { Button } from "@/components/ui/button";
 
@@ -26,9 +11,6 @@ import { highlighter } from "@/lib/highlighter";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { useFetch } from "@/lib/api";
 import { toast } from "sonner";
-import ConfirmationDialog from "@/components/dialogs/confirm-dialog";
-import CopyValueDialog from "@/components/dialogs/copy-value-dialog";
-import CopyButton from "@/components/copy-btn";
 import { m } from "@/lib/paraglide/messages";
 import { transformerNotationDiff, transformerRenderWhitespace } from "@shikijs/transformers";
 import { diffArrays } from "diff";
@@ -143,8 +125,6 @@ function ApplicationDetailsPage() {
 	const { data } = useFetch<Application>("/applications/" + id);
 
 	const [deploying, setDeploying] = useState(false);
-	const [webhookSecret, setWebhookSecret] = useState<string | null>(null);
-	const [webhookSecretOpen, setWebhookSecretOpen] = useState(false);
 
 	const handleDeploy = async () => {
 		setDeploying(true);
@@ -176,25 +156,6 @@ function ApplicationDetailsPage() {
 			transformers: [transformerNotationDiff(), transformerRenderWhitespace()],
 		});
 	}, [data?.previousComposeFile, data?.composeFile, theme]);
-
-	async function handleGenerateWebhook() {
-		try {
-			const result = await generateImageWebhook(id);
-			setWebhookSecret(result.secret);
-			setWebhookSecretOpen(true);
-		} catch (err) {
-			toast.error(err instanceof Error ? err.message : m.failedGenerateWebhook());
-		}
-	}
-
-	async function handleRevokeWebhook() {
-		try {
-			await revokeImageWebhook(id);
-			toast.success(m.webhookRevoked());
-		} catch (err) {
-			toast.error(err instanceof Error ? err.message : m.failedRevokeWebhook());
-		}
-	}
 
 	return (
 		<div className="space-y-6">
@@ -258,7 +219,6 @@ function ApplicationDetailsPage() {
 				<TabsList className="bg-muted">
 					<TabsTrigger value="manifest">{m.manifest()}</TabsTrigger>
 					<TabsTrigger value="diff">Diff</TabsTrigger>
-					<TabsTrigger value="webhook">{m.webhook()}</TabsTrigger>
 				</TabsList>
 
 				<TabsContent value="manifest" className="space-y-4">
@@ -275,64 +235,6 @@ function ApplicationDetailsPage() {
 							<div dangerouslySetInnerHTML={{ __html: diffHtml }} />
 						</div>
 					</div>
-				</TabsContent>
-
-				<TabsContent value="webhook" className="space-y-4">
-					<Card>
-						<CardHeader>
-							<CardTitle>
-								<div className="flex items-center gap-2">
-									<Webhook className="h-5 w-5" />
-									{m.imagePullWebhookSectionTitle()}
-								</div>
-							</CardTitle>
-						</CardHeader>
-						<CardContent className="space-y-4">
-							<p className="text-sm text-muted-foreground">{m.imagePullWebhookDescription()}</p>
-
-							{data?.imageWebhookEnabled && data.imageWebhookUrl ? (
-								<div className="space-y-1">
-									<p className="text-xs font-medium">{m.imagePullWebhookUrl()}</p>
-									<div className="flex items-center gap-1 rounded-md border bg-muted/50 px-3 py-1">
-										<code className="flex-1 truncate font-mono text-sm">
-											{data.imageWebhookUrl}
-										</code>
-										<CopyButton text={data.imageWebhookUrl} title={m.imagePullWebhookUrl()} />
-									</div>
-								</div>
-							) : null}
-
-							<div className="flex items-center gap-2">
-								<Button variant="outline" size="sm" onClick={handleGenerateWebhook}>
-									{data?.imageWebhookEnabled ? m.regenerateWebhook() : m.generateWebhook()}
-								</Button>
-								{data?.imageWebhookEnabled && (
-									<ConfirmationDialog
-										onConfirm={handleRevokeWebhook}
-										description={m.revokeWebhookConfirmDescription()}
-										triggerText={m.revokeWebhook()}
-										triggerProps={{ variant: "destructive", size: "sm" }}
-									/>
-								)}
-							</div>
-						</CardContent>
-					</Card>
-
-					<CopyValueDialog
-						open={webhookSecretOpen}
-						onOpenChange={(nextOpen) => {
-							setWebhookSecretOpen(nextOpen);
-							if (!nextOpen) {
-								setWebhookSecret(null);
-							}
-						}}
-						title={m.imagePullWebhookSecretModalTitle()}
-						description={m.imagePullWebhookSecretModalDescription()}
-						label={m.imagePullWebhookSecret()}
-						value={webhookSecret ?? ""}
-						inputId={`app-pull-webhook-secret-${id}`}
-						copyTitle={m.imagePullWebhookSecret()}
-					/>
 				</TabsContent>
 			</Tabs>
 		</div>
